refactor(test): Extract test_eof_handling.py
